// apps/api/src/generated/prisma/enums.ts
export const ProjectRole = {
  OWNER: 'OWNER',
  MANAGER: 'MANAGER',
  MEMBER: 'MEMBER',
  VIEWER: 'VIEWER'
} as const

export type ProjectRole = (typeof ProjectRole)[keyof typeof ProjectRole]